Create Bayesian Statistics
Bayesian methods underpin many Time Series Analysis and Machine Learning techniques, thus, an understanding of the "philosophy" of the Bayesian approach and how to apply it to real-world quantitative finance problems is essential.

Knowledge of the basic ideas of Bayesian Statistics, particularly Bayes’ Theorem (a.k.a Bayes’ Rule) is required. 

It is important to note that the Bayesian approach contrasts the more traditional Classical/Frequentist, approach to statistics.

Bayesian statistics - provides us with mathematical tools to rationally update our subjective beliefs in light of new data or evidence.

Frequentist statistics assumes that probabilities are the long-run frequency of random  events in repeated trials.